+++ src/arguments/index.ts
@@ -4,6 +4,7 @@ import { dofusJob } from './dofus/job';
 import { dofusServer } from './dofus/server';
 import { duration } from './time/duration';
 import { line } from './text/line';
+import { oneline } from './text/oneline';
 
 export const argumentTypes = {
   dofusCoordinates,
@@ -11,5 +12,6 @@ export const argumentTypes = {
   dofusJob,
   dofusServer,
   duration,
-  line
+  line,
+  oneline
 };