+++ src/models/index.ts
@@ -1,5 +1,6 @@
 export { almanaxModel as almanax } from '@/models/almanax';
 export { guildModel as guilds } from '@/models/guild';
+export { issueModel as issues } from '@/models/issue';
 export { logModel as logs } from '@/models/log';
 export { memberModel as members } from '@/models/member';
 export { pollModel as polls } from '@/models/poll';